backend/pyproject.toml
[tool.poetry]
name = "dailybrief-backend"
version = "0.1.0"
description = "DailyBrief ingestion, summarization, verification, ranking and delivery stack"
authors = ["DailyBrief Engineering <eng@dailybrief.ai>"]
license = "Proprietary"
readme = "README.md"
packages = [
  { include = "news_core", from = "libs" },
  { include = "vector_utils", from = "libs" },
  { include = "ranking_lib", from = "libs" }
]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = { extras = ["standard"], version = "^0.29.0" }
httpx = "^0.27.0"
feedparser = "^6.0.11"
beautifulsoup4 = "^4.12.3"
langdetect = "^1.0.9"
structlog = "^24.1.0"
loguru = "^0.7.2"
pydantic = "^2.7.0"
pydantic-settings = "^2.2.1"
sqlalchemy = "^2.0.29"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
prometheus-client = "^0.20.0"
opentelemetry-api = "^1.25.0"
opentelemetry-sdk = "^1.25.0"
opentelemetry-exporter-otlp = "^1.25.0"
opentelemetry-instrumentation-fastapi = "^0.46b0"
opentelemetry-instrumentation-httpx = "^0.46b0"
redis = "^5.0.4"
celery = "^5.4.0"
python-dotenv = "^1.0.1"
strawberry-graphql = "^0.209.6"
starlette = "^0.37.2"
networkx = "^3.3"
hdbscan = "^0.8.39"
scikit-learn = "^1.4.2"
numpy = "^1.26.4"
scipy = "^1.13.0"
sentence-transformers = "^2.6.0"
transformers = "^4.40.1"
boto3 = "^1.34.97"
pgvector = "^0.2.4"
rapidfuzz = "^3.6.1"
pyyaml = "^6.0.1"
typing-extensions = "^4.11.0"
aiohttp = "^3.9.5"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.6"
pytest-cov = "^5.0.0"
httpx = {version = "^0.27.0", extras = ["socks"]}
ruff = "^0.4.2"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
